Change internal FlatList to inner InnerList
