+++ src/collections-flat-array/Collections.FlatArray/FlatArray.T.InnerFlatList/InnerFlatList.cs
@@ -1,0 +1,11 @@
+﻿namespace System.Collections.Generic;
+
+partial struct FlatArray<T>
+{
+    private sealed partial class InnerFlatList : IList<T>, IReadOnlyList<T>
+    {
+        private readonly T[] items;
+
+        internal InnerFlatList(T[] items) => this.items = items;
+    }
+}
