--- src/collections-flat-array/Collections.FlatArray/FlatList.T/FlatList.Impl.Legacy.cs
@@ -1,12 +1,0 @@
-﻿namespace System.Collections.Generic;
-
-partial class FlatList<T>
-{
-    bool ICollection<T>.IsReadOnly => true;
-
-    T IList<T>.this[int index]
-    {
-        get => items[index];
-        set => throw InnerExceptionFactory.NotSupportedOnReadOnlyCollection();
-    }
-}
